moved auth into component and added login page
